correct logic of dependency installation in entrypoint.sh

+++ entrypoint.sh
@@ -1,13 +1,14 @@
 #!/bin/bash
 
-if [ ! -f dependencies.txt ] || ! cmp -s dependencies.txt /user_code/requirements.txt; then
-    pip3 install -r /user_code/requirements.txt
-    cp /user_code/requirements.txt dependencies.txt
-elif cmp -s dependencies.txt /user_code/requirements.txt; then
-    exec "$@"
+if [ -f dependencies.txt ] && cmp -s dependencies.txt /user_code/requirements.txt; then
+  exec "$@"
+elif [ -f dependencies.txt ] && ! cmp -s dependencies.txt /user_code/requirements.txt; then
+  echo "Rebuilding..."
+  pip3 uninstall -r /user_code/requirements.txt -y
+  pip3 install -r /user_code/requirements.txt
+  cp /user_code/requirements.txt dependencies.txt
+  exec "$@"
 else
-    pip3 uninstall -r /user_code/requirements.txt
-    pip3 install -r /user_code/requirements.txt
-    cp /user_code/requirements.txt dependencies.txt
-    exec "$@"
+  pip3 install -r /user_code/requirements.txt
+  cp /user_code/requirements.txt dependencies.txt
 fi
